Attempt to subdivide PyDictEntry blocks based on what the dictionaries are for
Initial usage of the tool on python showed "PyDictEntry" as the top category.

My hunch is that these are for the __dict__ of objects; perhaps we can
further sub-divide the report by the classes of these objects (thus
providing possible places where the __slots__ optimization can be used).

Add a level when categorizing; a higher level implies richer categorization.

Add a detail, so if we have a PyDictEntry we can optionally tell what the
dict is being used for.