Split up categorization into "domain", "kind" and "detail"
We were already loosely marking blocks with "category" and "detail",
now we have a Category class to capture this (actually a namedtuple
subclass, so instances are immutable), on the grounds that different
people will want to view and query for different aspects of memory
management.

Add __slots__ to Usage; it can't be a namedtuple since Usage instances
are mutable.

Update throughout (I hope) to use Category rather than strings