Mark the __dict__ (if any) of new-style classes with the classname
Add an optional "obj" field to Usage, so that we can associate PyObjectPtr
instances with them as appropriate.

Make categorize() take a Usage directly, so that we can assign to the obj
field

Add modified versions of parts of PyObjectPtr and PyTypeObjectPtr from my
pretty-printing hooks from Python upstream, creating PyTypeObjectPtr instances
when we detect a new-style CPython class in the inferior process.

Implement a PyObjectPtr.categorize_refs hook to allow these instances to mark
the buffers they reference with higher-level usage information.

Use this to mark the __dict__ in a new-style class with "CLASSNAME.__dict__"

Add a test for this to the test_python selftest